Optimized the parser logic
