+++ __tests__/parser/a.lowercase.spec.ts
@@ -1,5 +1,5 @@
 import {DateTimeFormatter} from "../../src/DateTimeFormatter";
-import FakeParser from "../../src/utils/FakeParser";
+import FakeParser from "../../src/classes/FakeParser";
 
 describe('correct', () => {
     for (const item of ['am', 'pm']) {
