+++ __tests__/parser/h.lowercase.spec.ts
@@ -1,4 +1,4 @@
-import FakeParser from "../../src/utils/FakeParser";
+import FakeParser from "../../src/classes/FakeParser";
 import {DateTimeFormatter} from "../../src/DateTimeFormatter";
 import {leadingZeroNumber} from "../../src/utils/leadingZeroNumber";
